requirements.txt
python-binance==1.0.19
requests==2.31.0
python-dotenv==1.0.0

# Note: tkinter is included with Python standard library
# If you're on Linux and tkinter is missing, install: sudo apt-get install python3-tk 